+++ data_adapter/core.py
@@ -1,6 +1,5 @@
 import json
 import pathlib
-from enum import IntEnum
 from typing import Union
 
 SCALAR_COLUMNS = {
@@ -25,10 +24,16 @@ TIMESERIES_COLUMNS = {
     "timeindex_resolution",
 }
 
-
-class DataType(IntEnum):
-    Scalar = 0
-    Timeseries = 1
+OEP_TO_FRICTIONLESS_CONVERSION = {
+    "int": "integer",
+    "bigint": "integer",
+    "text": "string",
+    "json": "object",
+    "decimal": "number",
+    "interval": "any",
+    "timestamp": "datetime",
+    "float": "number",
+}
 
 
 def get_metadata(metadata: Union[str, pathlib.Path, dict]):
@@ -38,9 +43,16 @@ def get_metadata(metadata: Union[str, pathlib.Path, dict]):
         return json.load(metadata_file)
 
 
-def get_data_type(metadata: Union[str, pathlib.Path, dict]):
-    metadata = get_metadata(metadata)
-    for field in metadata["resources"][0]["schema"]["fields"]:
-        if field["name"].startswith("timeindex"):
-            return DataType.Timeseries
-    return DataType.Scalar
+def reformat_oep_to_frictionless_schema(schema):
+    # Ignore other fields than 'fields' and 'primaryKey' (i.e. "foreignKeys")
+    fields = []
+    for field in schema["fields"]:
+        if "array" in field["type"]:
+            type_ = "array"
+        else:
+            type_ = OEP_TO_FRICTIONLESS_CONVERSION.get(field["type"], field["type"])
+        field_data = {"name": field["name"], "type": type_}
+        if field["type"] == "float":
+            field_data["floatNumber"] = "True"
+        fields.append(field_data)
+    return {"fields": fields, "primaryKey": schema["primaryKey"], "foreignkeys": []}