+++ data_adapter/preprocessing.py
@@ -1,101 +1,13 @@
-import json
 import pathlib
-from dataclasses import dataclass
 from typing import Dict, List
 
+import frictionless
 import pandas
 
-from data_adapter import settings, structure
-from data_adapter.core import SCALAR_COLUMNS, TIMESERIES_COLUMNS, DataType
+from data_adapter import collection, core, settings, structure
 
 
-@dataclass
-class Artifact:
-    collection: str
-    group: str
-    artifact: str
-    version: str
-    filename: str
-    subject: str
-    datatype: DataType
-
-    def path(self):
-        return (
-            settings.COLLECTIONS_DIR
-            / self.collection
-            / self.group
-            / self.artifact
-            / self.version
-            / self.filename
-        )
-
-
-def __get_collection_meta(collection: str) -> dict:
-    """
-    Returns collection meta file if present
-
-    Parameters
-    ----------
-    collection : str
-        Name of collection to get metadata from
-
-    Returns
-    -------
-    dict
-        Metadata for given collection
-
-    Raises
-    ------
-    FileNotFoundError
-        if collection metadata cannot be found in given collection folder
-    """
-    collection_folder = pathlib.Path(settings.COLLECTIONS_DIR) / collection
-    collection_meta_file = collection_folder / settings.COLLECTION_JSON
-    if not collection_meta_file.exists():
-        raise FileNotFoundError(
-            f"Could not find collection meta ('{settings.COLLECTION_JSON}') in collection folder '{collection_folder}'."
-        )
-    with open(collection_meta_file, "r", encoding="utf-8") as meta_file:
-        return json.load(meta_file)
-
-
-def __get_artifacts_for_process(collection: str, process: str) -> List[Artifact]:
-    """
-    Returns list of artifacts belonging to given process (subject)
-
-    Parameters
-    ----------
-    collection: str
-        Collection name
-    process : str
-        Name of process to search collection metadata for
-
-    Returns
-    -------
-    List[ArtifactPath]
-        List of artifacts in collection belonging to given process
-    """
-    collection_meta = __get_collection_meta(collection)
-    artifacts = []
-    for group in collection_meta:
-        for artifact, artifact_info in collection_meta[group].items():
-            if artifact_info["subject"] == process:
-                filename = f"{artifact}.csv"
-                artifacts.append(
-                    Artifact(
-                        collection,
-                        group,
-                        artifact,
-                        artifact_info["latest_version"],
-                        filename,
-                        subject=process,
-                        datatype=DataType(artifact_info["datatype"]),
-                    )
-                )
-    return artifacts
-
-
-def __get_df_from_artifact(artifact: Artifact, *parameters: List[str]):
+def __get_df_from_artifact(artifact: collection.Artifact, *parameters: List[str]):
     """
     Returns DataFrame from given artifact.
 
@@ -113,20 +25,31 @@ def __get_df_from_artifact(artifact: Artifact, *parameters: List[str]):
     -------
     pandas.DataFrame
     """
-    df = pandas.read_csv(artifact.path())
+    metadata = collection.get_metadata_from_artifact(artifact)
+    fl_table_schema = core.reformat_oep_to_frictionless_schema(
+        metadata["resources"][0]["schema"]
+    )
+    resource = frictionless.Resource(
+        name=metadata["name"],
+        profile="tabular-data-resource",
+        source=artifact.path() / f"{artifact.filename}.csv",
+        schema=fl_table_schema,
+        format="csv",
+    )
+    df = resource.to_pandas()
     if len(parameters) == 0:
         return df
     columns = (
-        set(SCALAR_COLUMNS)
-        if artifact.datatype is DataType.Scalar
-        else set(TIMESERIES_COLUMNS)
+        set(core.SCALAR_COLUMNS)
+        if artifact.datatype is collection.DataType.Scalar
+        else set(core.TIMESERIES_COLUMNS)
     )
     columns.update(set(parameters))
     drop_columns = set(df.columns).difference(columns)
     return df.drop(drop_columns, axis=1)
 
 
-def get_process_df(collection: str, process: str) -> Dict[str, pandas.DataFrame]:
+def get_process_df(collection_name: str, process: str) -> Dict[str, pandas.DataFrame]:
     """
     Loads data for given process from collection as pandas.DataFrame
 
@@ -134,7 +57,7 @@ def get_process_df(collection: str, process: str) -> Dict[str, pandas.DataFrame]
 
     Parameters
     ----------
-    collection : str
+    collection_name : str
         Name of collection to get data from
     process : str
         Name of process (from subject)
@@ -151,17 +74,17 @@ def get_process_df(collection: str, process: str) -> Dict[str, pandas.DataFrame]
     StructureError
         if additional parameters of process are related to multiple subjects
     """
-    collection_folder = pathlib.Path(settings.COLLECTIONS_DIR) / collection
+    collection_folder = pathlib.Path(settings.COLLECTIONS_DIR) / collection_name
     if not collection_folder.exists():
         raise FileNotFoundError(
-            f"Could not find {collection=} in collection folder '{collection_folder}'."
+            f"Could not find {collection_name=} in collection folder '{collection_folder}'."
         )
-    artifacts = __get_artifacts_for_process(collection, process)
+    artifacts = collection.get_artifacts_from_collection(collection_name, process)
     data = {}
     for artifact in artifacts:
         data[artifact.artifact] = __get_df_from_artifact(artifact)
     for subject, parameters in structure.get_additional_parameters(process).items():
-        artifacts = __get_artifacts_for_process(collection, subject)
+        artifacts = collection.get_artifacts_from_collection(collection_name, subject)
         if len(artifacts) > 1:
             raise structure.StructureError(
                 f"Additional parameter for process '{process}' "